requirements.txt
python-dotenv
google-generativeai
psycopg2-binary
pypdf
python-docx
tqdm
pysbd
